Made contexts be current on the expected threads on start. Allowed setting the current context to null. Made the glfw window pointer public so you don't need to subclass NativeWindow or GameWindow to gain access. Removed the separate class used to load glfw, made it a local function instead. Fixed a lot of .net versions in the different projects (removing .net framework stuff).
