+++ src/OpenTK.Windowing.GraphicsLibraryFramework/OpenTK.Windowing.GraphicsLibraryFramework.csproj
@@ -4,6 +4,7 @@
       <TargetFrameworks>netcoreapp3.1</TargetFrameworks>
       <RootNamespace>OpenTK.Windowing.GraphicsLibraryFramework</RootNamespace>
       <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+      <LangVersion>8.0</LangVersion>
     </PropertyGroup>
     
     <ItemGroup>
